% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_DeGy_columns.R
\name{add_DeGy_columns}
\alias{add_DeGy_columns}
\title{Add or Update Dose Columns in Excel Data and Save, instead of doing it manually.}
\usage{
add_DeGy_columns(file_path, dose_rate, dose_rate_err)
}
\arguments{
\item{file_path}{Path to the input Excel file.}

\item{dose_rate}{Environmental dose rate in Gy/ka.}

\item{dose_rate_err}{Uncertainty in the dose rate in Gy/ka.}
}
\value{
A data frame with updated dose columns.
}
\description{
This function reads an Excel file, calculates the equivalent dose and uncertainty in Gy using the provided dose rate and uncertainty, and updates or creates the columns: "De Gy", "De err Gy", "D0 Gy", "D0 err Gy". The updated Excel file is saved.
}
\examples{
\dontrun{
updated_data <- add_missing_columns("data.xlsx", dose_rate = 2.5, dose_rate_err = 0.1)
}
}
