% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_ages.R
\name{adm_ages}
\alias{adm_ages}
\title{adm_ages
Takes De Gy values from list and calls to the function calc_AverageDose.
Saves results as a summary table as a dataframe and has options to save the results as a csv file.}
\usage{
adm_ages(datasets, sigma, path_for_save = NULL, plots = FALSE)
}
\arguments{
\item{datasets}{A list containing OSL data, which must include columns labelled De Gy, De errGy, D0 Gy and D0 err Gy}

\item{sigma}{Value of overdispersion, taken from a dose recovery measurement.}

\item{path_for_save}{Directory for where you want to save the file.}

\item{plots}{TRUE or FALSE (default). If TRUE, standard plots from the calc_AverageDose function will be plotted.}
}
\value{
adm_outputs
}
\description{
adm_ages
Takes De Gy values from list and calls to the function calc_AverageDose.
Saves results as a summary table as a dataframe and has options to save the results as a csv file.
}
