% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_ages.R
\name{cam_ages}
\alias{cam_ages}
\title{Title}
\usage{
cam_ages(datasets, path_for_save = NULL, plots = FALSE)
}
\arguments{
\item{datasets}{A list containing OSL data, which must include columns labelled De Gy, De errGy, D0 Gy and D0 err Gy}

\item{path_for_save}{Directory for where you want to save the file.}

\item{plots}{TRUE or FALSE (default). If TRUE, standard plots from the calc_AverageDose function will be plotted.}
}
\value{
cam_output. COntains table of CAM ages, Over dispersion
}
\description{
Title
}
