% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_OSL_grains.R
\name{filter_OSL_grains}
\alias{filter_OSL_grains}
\title{filter_osl_grains}
\usage{
filter_OSL_grains(
  data,
  filter_options = "all",
  d0filt_limit = NULL,
  plots = TRUE
)
}
\arguments{
\item{data}{Data frame with OSL data, which must contain columns labelled De Gy, De errGy, D0 Gy and D0 err Gy}

\item{filter_options}{list of strings to state what type of filtering you want for your grains. #   "positive" WHich only removes negative grains
"feldsparfilt" WHich removes negative De avlues, and then only keeps grains with IRSL depletion ratio 0.9<>1
"d0filt" after filtering for feldspars, this removes grains with D0 values less than the mean De (Default), or less than the value stated in d0filt
"recuperation" after filtering for feldspars,Removes grains with recuperation values < 0.1
"zerodoserem" after filtering for feldspars, removes grains with doses < 1 Gy
"nanrem" after filtering for feldspars, it removes  grains with De as NAN}

\item{d0filt_limit}{Optional. Numeric value to be used as a lower threshold for D0 filtering.}

\item{plots}{TRUE (default) or FALSE. If TRUE, histogram and dnesity plots of De values are show, as well as De Gy VS. DE error Gy plots, and De Gy VS> D0 Gy plots. .}
}
\value{
Returns list object with filtered data sets. COlumn names are the same as the strings used for selecting filtering types
}
\description{
filter_osl_grains
}
