% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain_summary_statistics.R
\name{grain_summary_statistics}
\alias{grain_summary_statistics}
\title{Title}
\usage{
grain_summary_statistics(datasets, path_for_save = NULL)
}
\arguments{
\item{datasets}{Contains list of data sets, each with different names.}

\item{path_for_save}{Optional. String with path where you want to save your results.}
}
\value{
Returns a table with all summary statistics in. If Path for save is supplied, this table will be saved as a CSV file.
}
\description{
Title
}
