% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain_summary_stats.R
\name{grain_summary_stats}
\alias{grain_summary_stats}
\title{This code calculates different summary statistics for data sets which have been filtered by different means. You must input lists of De values.}
\usage{
grain_summary_stats(datasets, samplename, path_for_save = NULL)
}
\arguments{
\item{datasets}{Contains list of data sets, each with different names.}

\item{samplename}{String with your sample name.}

\item{path_for_save}{Optional. String with path where you want to save your results.}
}
\value{
Returns a table with all summary statistics in. If Path for save is supplied, this table will be saved as a CSV file.
}
\description{
This code calculates different summary statistics for data sets which have been filtered by different means. You must input lists of De values.
}
