% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load Example Data}
\usage{
load_example_data()
}
\value{
A data frame containing the example data.
}
\description{
Loads the example dataset from the package's extdata.
}
