% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_discpos_files.R
\name{save_discpos_files}
\alias{save_discpos_files}
\title{Title}
\usage{
save_discpos_files(
  path,
  data,
  discpos_to_save = NULL,
  discpos_to_use = NULL,
  env_dose_rates = NULL,
  reader_dose_rates = NULL
)
}
\arguments{
\item{path}{Directory path for where you whant to save your data.}

\item{data}{List of OSL data.}

\item{discpos_to_save}{Optional. List of strings of which DiscPos files you want to save. Each data set in the list will be saved if no data set is specified.}

\item{discpos_to_use}{Optiona. State here which data set in particular you want to start using, and this DiscPos file will be labelled as DiscPos.csv and will be directly usable for BayLum.}

\item{env_dose_rates}{Optional.}

\item{reader_dose_rates}{Optional.}
}
\value{
Does not return any value to R.
}
\description{
Title
}
