% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_discpos_files.R
\name{save_discpos_files}
\alias{save_discpos_files}
\title{Options to save CSV files consisting of 2 columns Disc;Pos
Ready for use with BayLum.}
\usage{
save_discpos_files(
  path,
  data,
  discpos_to_use = NULL,
  env_dose_rates = NULL,
  reader_dose_rates = NULL
)
}
\arguments{
\item{path}{Directory path for where you whant to load your data from or save your data.}

\item{data}{List of OSL data.}

\item{discpos_to_use}{Optional. State here which data set in particular you want to start using, and this DiscPos file will be labelled as DiscPos.csv and will be directly usable for BayLum.}

\item{env_dose_rates}{Optional.}

\item{reader_dose_rates}{Optional.}
}
\value{
Returns files in your set path where lists of disc;pos for desired grains are saved as a CSV file.
}
\description{
Options to save CSV files consisting of 2 columns Disc;Pos
Ready for use with BayLum.
}
